Change Bind implementation to use `flatMap` instead of `flatMapLatest`
This was changed a while back to `flatMapLatest` with the idea that in the
browser environment it's a common bug to use plain `flatMap`. In network
programming `flatMap` is what you almost always want. Many generic Monad
functions tend to have a more common use case with `flatMap`, like `join`.

Do-notation is now rebindable in PureScript. So when `flatMapLatest` is needed
in a do-block, you can just rebind `>>=`.